fix: properly close broadcast channel to prevent memory leak
With a large number of requesting clients, the software keeps open an evenly large number of client structs and broadcastChannels filled with certificates.

This is due to the broadcastHandler goroutine waiting for new elements to arrive in the channel. Since that does not happen, the client struct will not be freed and thus heap memory keeps piling up.

In my case the tool used up ~3-4 GB of memory after 228 days of runtime with a single client that connected and disconnected frequently.