fix: ip whitelists for the websocket server
Previously, the whitelist code had several issues. It always read the Prometheus whitelist which lead to the webserver whitelist not being applied.

Also, using a whitelist on the webserver lead to a crash (as seen in #33). That was due to the wrong order of chi commands being executed. It is not possible to define middlewares after we defined routes.

Thirdly, the code for parsing single IPs was broken. The IP was not stored in the proper variable and neither did we skip the entry if it couldn't get parsed.

Fixes #33