feat: start new ct-watchers as new ct logs become available
fixes #42

We also needed a way to properly prevent the entries of the metrics map from being reset to 0. That's why I added the Init() method to the LogMetrics. It only sets the value of the operator/url combination to 0 if it doesn't exist before.

If it already exists, it will be left alone.