fix: remove unnecessary call to updateLogs
This was executed when starting up the tool. Since the updateLogs gets called just a few lines before, it doesn't make sense to call it yet another time.

It needs to be called once. Otherwise there won't be any workers which causes the waitgroup to not wait, which in turn causes the certChan to be closed which then causes the application to quit.

Calling updateLogs in the watchNewLogs method causes undefined behavior since watchNewLogs is called in a goroutine and in most cases the primary execution flow will reach the wg.Wait() call before new workers got added.